# docs/swagger.yaml
basePath: /users-segmentation
definitions:
  avito.AlteredUserSegments:
    properties:
      add:
        items:
          type: string
        type: array
      delete:
        items:
          type: string
        type: array
      id:
        type: string
    required:
    - id
    type: object
  avito.CustomDate:
    properties:
      date:
        type: string
    required:
    - date
    type: object
  avito.Segment:
    properties:
      name:
        type: string
    required:
    - name
    type: object
  avito.User:
    properties:
      name:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - name
    - password
    - username
    type: object
  handler.errorResponse:
    properties:
      message:
        type: string
    type: object
host: localhost:8000
info:
  contact: {}
  description: API Service for user's segments
  title: User Segmentation API
  version: "1.0"
paths:
  /segment:
    delete:
      consumes:
      - application/json
      description: delete segment by name
      operationId: delete-segment
      parameters:
      - description: segment name
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/avito.Segment'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: DeleteSegment
      tags:
      - segment
    post:
      consumes:
      - application/json
      description: create a new segment
      operationId: create-segment
      parameters:
      - description: segment info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/avito.Segment'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: CreateSegment
      tags:
      - segment
  /user:
    post:
      consumes:
      - application/json
      description: create a new user
      operationId: create-user
      parameters:
      - description: user info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/avito.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: CreateUser
      tags:
      - user
  /user/{id}:
    delete:
      consumes:
      - application/json
      description: delete user by id
      operationId: delete-user
      parameters:
      - description: the UUID of a user
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: DeleteUser
      tags:
      - user
  /users-segments:
    post:
      consumes:
      - application/json
      description: update user's segments
      operationId: update-user-segments
      parameters:
      - description: user_id and segments to add/delete
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/avito.AlteredUserSegments'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: UpdateUserSegments
      tags:
      - user_segments
  /users-segments/{id}:
    get:
      consumes:
      - application/json
      description: get user's segments
      operationId: get-user-segments
      parameters:
      - description: the UUID of a user
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: GetUserSegments
      tags:
      - user_segments
  /users-segments/logs:
    post:
      consumes:
      - application/json
      description: get user's segments logs
      operationId: get-user-segments-logs
      parameters:
      - description: year and month
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/avito.CustomDate'
      produces:
      - text/csv
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: GetUserSegmentsLogs
      tags:
      - user_segments
swagger: "2.0"